+++ src/main/java/de/uib/utils/Utils.java
@@ -36,6 +36,8 @@ import javax.swing.JScrollPane;
 import javax.swing.JTextArea;
 import javax.swing.SwingUtilities;
 
+import com.formdev.flatlaf.extras.FlatSVGIcon;
+
 import de.uib.configed.Configed;
 import de.uib.configed.ConfigedMain;
 import de.uib.configed.Globals;
@@ -482,5 +484,54 @@ public final class Utils {
 			}
 		};
 		optionPane.addHierarchyListener(listener);
+	}
+
+	public static Boolean toBoolean(Object obj) {
+		if (obj instanceof Boolean bool) {
+			return bool;
+		}
+		if (obj instanceof String) {
+			return Boolean.valueOf(obj.toString());
+		}
+		return false;
+	}
+
+	public static FlatSVGIcon determineIconBasedOnPlatform(String platform, int size) {
+		FlatSVGIcon icon = null;
+		if ("macos".equals(platform)) {
+			icon = Icons.getThemeSVGRepoIcon("macos", size);
+		} else if ("windows".equals(platform)) {
+			icon = Icons.getThemeSVGRepoIcon("windows", size);
+		} else if ("linux".equals(platform)) {
+			//icon = Icons.getThemeIntellijIcon("linux", size);
+			icon = Icons.getThemeSVGRepoIcon("linux", size);
+		} else {
+			icon = Icons.getThemeIntellijIcon("questionMark", size);
+			//icon = null;
+		}
+		return icon;
+	}
+
+	public static FlatSVGIcon determineIconBasedOnDeviceType(String value) {
+		return determineIconBasedOnDeviceType(value, 16);
+	}
+
+	public static FlatSVGIcon determineIconBasedOnDeviceType(String value, int size) {
+		FlatSVGIcon icon = null;
+		if ("notebook".equals(value)) {
+			icon = Icons.getThemeSVGRepoIcon("laptop", size);
+		} else if ("desktop".equals(value)) {
+			icon = Icons.getThemeSVGRepoIcon("desktop", size);
+		} else if ("virtual_machine".equals(value)) {
+			icon = Icons.getThemeSVGRepoIcon("virtualMachine", size);
+		} else if ("convertible".equals(value)) {
+			icon = Icons.getThemeSVGRepoIcon("convertible", size);
+		} else if ("other".equals(value)) {
+			icon = Icons.getThemeIntellijIcon("questionMark", size);
+		} else {
+			icon = Icons.getThemeIntellijIcon("questionMark", size);
+			//icon = null;
+		}
+		return icon;
 	}
 }